Seperated the graph component using useEffect
